added salt for user's key, changed cryptoservice's api
