+++ Maelstorm/Dtos/AuthenticationResultDTO.cs
@@ -8,6 +8,7 @@ namespace Maelstorm.Dtos
     public class AuthenticationResultDTO
     {
         public string IVBase64 { get; set; }
+        public string KeySaltBase64 { get; set; }
         public string EncryptedPrivateKey { get; set; }
         public TokensDTO Tokens { get; set; }
     }
