+++ Maelstorm/Services/Interfaces/ICryptographyService.cs
@@ -7,9 +7,10 @@ namespace Maelstorm.Services.Interfaces
 {
     public interface ICryptographyService
     {
-        byte[] AesEncryptBytes(byte[] bytes, string key, byte[] iv, int keySize = 128);
-        string GetRandomString();
-        string GeneratePasswordHash(string password, string salt);
+        byte[] AesEncryptBytes(byte[] bytes, byte[] key, byte[] iv, int keySize = 128);
+        byte[] GenerateSalt(int size = 32);
+        string GetRandomBase64String(int byteArraySize = 32);
+        byte[] Pbkdf2(string password, byte[] salt, int numBytes = 32);
         byte[] GenerateIV();
     }
 }
