+++ Maelstorm/Views/Message/Conversations.cshtml
@@ -213,6 +213,7 @@
 <script src="~/lib/fingerprint/fingerprint2.js"></script>
 <script src="~/lib/jquery/dist/jquery.min.js"></script>
 <script src="~/lib/signalr/dist/browser/signalr.js"></script>
+<script src="~/js/modules/formModule.js"></script>
 <script src="~/js/modules/accountModule.js"></script>
 <script src="~/js/modules/encodingModule.js"></script>
 <script src="~/js/modules/cryptoModule.js"></script>
